move JobScriptFiles to a dedicated file
